+++ util/coadd.c
@@ -158,7 +158,7 @@ number* coadd_get_snapshot(coadd_t* co, number* outimg,
                            number badpix) {
     int i;
     if (!outimg)
-	outimg = calloc(co->W * co->H, sizeof(number));
+	outimg = calloc((size_t)co->W * (size_t)co->H, sizeof(number));
 
     for (i=0; i<(co->W * co->H); i++) {
         if (co->weight[i] == 0)
@@ -190,7 +190,7 @@ void coadd_free(coadd_t* ca) {
 number* coadd_create_weight_image_from_range(const number* img, int W, int H,
                                              number lowval, number highval) {
     int i;
-    number* weight = malloc(W*H*sizeof(number));
+    number* weight = malloc((size_t)W*(size_t)H*sizeof(number));
     for (i=0; i<(W*H); i++) {
         if (img[i] <= lowval)
             weight[i] = 0;
